Improve readability

+++ src/identity/mod.rs
@@ -4,7 +4,6 @@ use crate::identity::storage::{
     InMemoryPenaltyStorage, InMemoryProofStorage, InMemoryVouchStorage, PenaltyStorage,
     ProofStorage, VouchStorage,
 };
-// Database storage implementations are available in `storage_db` module
 
 mod decay;
 pub mod error;
@@ -19,8 +18,8 @@ mod tree_walk;
 pub mod vouch;
 
 pub type UserAddress = String;
-pub type ProofId = u128;
-pub type IdtAmount = u128;
+pub type ProofId = u64;
+pub type IdtAmount = u64;
 
 #[derive(Clone)]
 pub struct ModeratorProof {
